Decouple search logic in a new class
